code/google_functions/update_covid_data/requirements.txt
pandas
requests
google-cloud-bigquery==1.24.0
google-cloud-storage==1.28.0

